+++ hass-actronque/hass-actronque/Que.Constants.cs
@@ -119,13 +119,13 @@ namespace HMX.HASSActronQue
 		}
 
 		// Timer/default values (seconds)
-		private static int _iCancellationTime = 15; // Seconds
+		private static int _iCancellationTime = 45; // Seconds
 		private static int _iPollInterval = 30; // Seconds
 		private static int _iPollIntervalOn = 60; // Seconds
 		private static int _iPollIntervalOff = 600; // Seconds		
 		private static int _iAuthenticationInterval = 60; // Seconds
 		private static int _iQueueInterval = 4; // Seconds
-		private static int _iCommandExpiry = 12; // Seconds
+		private static int _iCommandExpiry = 60; // Seconds
 		private static int _iPostCommandSleepTimerNeoNoEventsMode = 10; // Seconds
 		private static int _iTokenRefreshBufferSeconds = 300; // 5 minutes - refresh tokens this many seconds before expiry